@@ -1,136 +1,82 @@
 {
-  "name": "dashtar-admin",
-  "version": "0.1.0",
-  "description": "this is Dashtar-admin built with react.js and tailwindcss",
-  "private": true,
-  "dependencies": {
-    "@cloudinary/react": "^1.11.2",
-    "@cloudinary/url-gen": "^1.9.2",
-    "@headlessui/react": "^1.4.1",
-    "@pathofdev/react-tag-input": "^1.0.7",
-    "@react-pdf/font": "2.2.0",
-    "@react-pdf/renderer": "2.1.0",
-    "@shopify/shopify-api": "^7.3.0",
-    "@shopify/shopify-app-express": "^2.1.2",
-    "@shopify/shopify-app-session-storage-sqlite": "^1.2.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/draft-js": "^0.11.9",
-    "@windmill/react-ui": "^0.5.1",
-    "ajv": "^8.11.2",
-    "aws-sdk": "^2.1395.0",
-    "axios": "^0.24.0",
-    "babel-jest": "^26.6.0",
-    "chart.js": "^3.7.1",
-    "classnames": "2.2.6",
-    "cloudinary": "^1.35.0",
-    "combinate": "^1.1.11",
-    "csvtojson": "^2.0.10",
-    "dayjs": "^1.10.7",
-    "dotenv": "^16.0.1",
-    "draft-js": "^0.11.7",
-    "export-from-json": "^1.7.0",
-    "http-proxy-middleware": "^2.0.6",
-    "i18next": "^21.8.5",
-    "i18next-browser-languagedetector": "^6.1.4",
-    "immutability-helper": "^3.1.1",
-    "js-cookie": "^3.0.1",
-    "mock-aws-s3": "^4.0.2",
-    "multiselect-react-dropdown": "^2.0.21",
-    "nodemon": "^2.0.22",
-    "rc-drawer": "^4.4.2",
-    "rc-tree": "^5.6.5",
-    "react": "^17.0.2",
-    "react-beforeunload": "^2.5.3",
-    "react-chartjs-2": "^3.3.0",
-    "react-custom-scrollbars-2": "^4.4.0",
-    "react-dnd": "^16.0.1",
-    "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
-    "react-draft-wysiwyg": "^1.15.0",
-    "react-dropzone": "^11.4.2",
-    "react-flags-select": "^2.2.3",
-    "react-google-login": "^5.2.2",
-    "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.18.0",
-    "react-i18next": "^11.16.9",
-    "react-icons": "^4.3.1",
-    "react-loading-skeleton": "^3.2.0",
-    "react-multi-select-component": "^4.2.3",
-    "react-redux": "^8.0.2",
-    "react-responsive-modal": "^6.2.0",
-    "react-router-dom": "^5.3.0",
-    "react-search-autocomplete": "^6.1.2",
-    "react-simple-keyboard": "^3.4.197",
-    "react-spinners": "^0.11.0",
-    "react-switch": "^7.0.0",
-    "react-tabs": "^4.2.1",
-    "react-tag-input-component": "^1.1.1",
-    "react-to-print": "^2.14.7",
-    "react-toastify": "^8.0.3",
-    "react-tooltip": "^4.2.21",
-    "react-transition-group": "^4.4.2",
-    "react-use-cart": "^1.13.0",
-    "redux": "^4.2.0",
-    "redux-thunk": "^2.4.1",
-    "socket.io-client": "^4.5.3",
-    "sweetalert": "^2.1.2",
-    "web-vitals": "^1.1.2",
-    "xlsx": "^0.18.5"
-  },
-  "resolutions": {
-    "@react-pdf/font": "2.2.0",
-    "react-scripts/**/react-error-overlay": "6.0.9"
-  },
+  "name": "@minimal/material-kit-react",
+  "author": "minimals.cc",
+  "licence": "MIT",
+  "version": "1.7.0",
+  "private": false,
   "scripts": {
-    "tailwind:dev": "tailwindcss build src/assets/css/tailwind.css -o src/assets/css/tailwind.output.css",
-    "tailwind:build": "cross-env NODE_ENV=production postcss src/assets/css/tailwind.css -o src/assets/css/tailwind.output.css",
-    "prestart": "npm run tailwind:dev",
     "start": "react-scripts start",
-    "prebuild": "npm run tailwind:build",
-    "build": "react-scripts build --stats",
-    "analyze": "webpack-bundle-analyzer build/bundle-stats.json",
-    "cz": "git-cz",
-    "release": "release-it",
+    "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "devDependencies": {
-    "autoprefixer": "^9.8.8",
-    "cross-env": "7.0.2",
-    "cssnano": "4.1.10",
-    "postcss-cli": "7.1.1",
-    "react-error-overlay": "6.0.9",
-    "react-scripts": "^4.0.3",
-    "release-it": "13.6.4",
-    "tailwindcss": "^1.9.6",
-    "webpack-bundle-analyzer": "^4.8.0"
-  },
-  "release-it": {
-    "github": {
-      "release": true
-    },
-    "npm": {
-      "publish": false
-    }
+    "eject": "react-scripts eject",
+    "lint": "eslint --ext .js,.jsx ./src",
+    "lint:fix": "eslint --fix --ext .js,.jsx ./src",
+    "clear-all": "rm -rf build node_modules",
+    "re-start": "rm -rf build node_modules && yarn install && yarn start",
+    "re-build": "rm -rf build node_modules && yarn install && yarn build"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app"
+    ]
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-react"
+    ]
   },
   "browserslist": {
     "production": [
-      ">0.3%",
-      "not ie 11",
+      ">0.2%",
       "not dead",
       "not op_mini all"
     ],
     "development": [
-      ">0.3%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
     ]
   },
-  "proxy": "https://mtreports.mandasadevelopment.com:80"
+  "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@faker-js/faker": "^7.6.0",
+    "@iconify/react": "^4.1.0",
+    "@mui/lab": "^5.0.0-alpha.103",
+    "@mui/material": "^5.11.10",
+    "apexcharts": "^3.37.0",
+    "change-case": "^4.1.2",
+    "date-fns": "^2.29.3",
+    "history": "^5.3.0",
+    "lodash": "^4.17.21",
+    "numeral": "^2.0.6",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-apexcharts": "^1.4.0",
+    "react-dom": "^18.2.0",
+    "react-helmet-async": "^1.3.0",
+    "react-hook-form": "^7.43.1",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "^5.0.1",
+    "simplebar-react": "^3.2.1",
+    "web-vitals": "^3.1.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.21.0",
+    "@babel/eslint-parser": "^7.19.1",
+    "eslint": "^8.34.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.8.4"
+  },
+  "overrides": {
+    "nth-check": "2.1.1"
+  }
 }
